+++ render-site/src/main/java/com/example/damage/DamageFormula.java
@@ -6,68 +6,67 @@ public class DamageFormula {
     public static final double DEFAULT_STAB = 1.5;
 
     // Core base multiplier pieces (kept consistent everywhere)
-    private static double core(int level, int power, int attack, int defense) {
+    private static double core(int level, int power, float attack, float defense) {
         // Use doubles to avoid integer-division truncation
         return level * (power / 10.0) * ((double) attack / Math.max(defense, 1));
     }
 
-    // --- Formula helpers (these mirror your girlfriend's naming) ---
 
-    public static int formulaSESTABNC(int level, int power, int attack, int defense,
+    public static int formulaSESTABNC(int level, int power, float attack, float defense,
                                      double superEffectiveLevel, double stabLevel) {
         return asInt(core(level, power, attack, defense) * stabLevel * superEffectiveLevel);
     }
 
-    public static int formulaSESTABCrit(int level, int power, int attack, int defense,
+    public static int formulaSESTABCrit(int level, int power, float attack, float defense,
                                         double superEffectiveLevel, double stabLevel) {
         return asInt(core(level, power, attack, defense) * stabLevel * superEffectiveLevel * 2.0);
     }
 
-    public static int formulaSENC(int level, int power, int attack, int defense,
+    public static int formulaSENC(int level, int power, float attack, float defense,
                                  double superEffectiveLevel) {
         return asInt(core(level, power, attack, defense) * superEffectiveLevel);
     }
 
-    public static int formulaSECrit(int level, int power, int attack, int defense,
+    public static int formulaSECrit(int level, int power, float attack, float defense,
                                    double superEffectiveLevel) {
         return asInt(core(level, power, attack, defense) * superEffectiveLevel * 2.0);
     }
 
-    public static int formulaNVESTABNC(int level, int power, int attack, int defense,
+    public static int formulaNVESTABNC(int level, int power, float attack, float defense,
                                        double notVeryEffectiveLevel, double stabLevel) {
         return asInt(core(level, power, attack, defense) * stabLevel * notVeryEffectiveLevel);
     }
 
-    public static int formulaNVESTABCrit(int level, int power, int attack, int defense,
+    public static int formulaNVESTABCrit(int level, int power, float attack, float defense,
                                          double notVeryEffectiveLevel, double stabLevel) {
         return asInt(core(level, power, attack, defense) * stabLevel * notVeryEffectiveLevel * 2.0);
     }
 
-    public static int formulaNVENC(int level, int power, int attack, int defense,
+    public static int formulaNVENC(int level, int power, float attack, float defense,
                                    double notVeryEffectiveLevel) {
         return asInt(core(level, power, attack, defense) * notVeryEffectiveLevel);
     }
 
-    public static int formulaNVECrit(int level, int power, int attack, int defense,
+    public static int formulaNVECrit(int level, int power, float attack, float defense,
                                      double notVeryEffectiveLevel) {
         return asInt(core(level, power, attack, defense) * notVeryEffectiveLevel * 2.0);
     }
 
-    public static int formulaSTABNC(int level, int power, int attack, int defense,
+    public static int formulaSTABNC(int level, int power, float attack, float defense,
                                     double stabLevel) {
         return asInt(core(level, power, attack, defense) * stabLevel);
     }
 
-    public static int formulaSTABCrit(int level, int power, int attack, int defense,
+    public static int formulaSTABCrit(int level, int power, float attack, float defense,
                                       double stabLevel) {
         return asInt(core(level, power, attack, defense) * stabLevel * 2.0);
     }
 
-    public static int formulaBase(int level, int power, int attack, int defense) {
+    public static int formulaBase(int level, int power, float attack, float defense) {
         return asInt(core(level, power, attack, defense));
     }
 
-    public static int formulaBaseCrit(int level, int power, int attack, int defense) {
+    public static int formulaBaseCrit(int level, int power, float attack, float defense) {
         return asInt(core(level, power, attack, defense) * 2.0);
     }
 
@@ -76,8 +75,8 @@ public class DamageFormula {
     public static int doMath(
             int level,
             int power,
-            int attack,
-            int defense,
+            float attack,
+            float defense,
             boolean isSuperEffective,
             double superEffectiveLevel,
             boolean isNotVeryEffective,
